feat(dashboard): dashboard layout
