+++ react-app/src/modules/dashboard/Dashboard.tsx
@@ -1,65 +1,81 @@
 import { MainRoot } from "@/components/navigation/MainRoot.tsx";
 import { MapContainer, Marker, Popup, TileLayer } from "react-leaflet";
-import { Card, CardHeader, CardTitle } from "@/components/ui/card.tsx";
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card.tsx";
 import { useCurrentUser } from "@/common/hooks/useCurrentUser.ts";
 import { useEffect } from "react";
-import { TbCalendarClock, TbCalendarRepeat } from "react-icons/tb";
-import { MdOutlineLaptop } from "react-icons/md";
+import { FieldRow } from "@/components/fieldRow.tsx";
+import { Button } from "@/components/ui/button.tsx";
+import { BsFillInfoSquareFill } from "react-icons/bs";
+import { useNavigate } from "react-router-dom";
+import { PlusIcon } from "@radix-ui/react-icons";
+import { MdHomeRepairService } from "react-icons/md";
+import { GrGroup } from "react-icons/gr";
+import { NotificationsCard } from "@/modules/profile/components/notificationsCard.tsx";
 
 export const Dashboard = () => {
   const { currentUser, refreshCurrentUser } = useCurrentUser();
+  const navigate = useNavigate();
 
   useEffect(() => {
     refreshCurrentUser();
   }, []);
 
+  const handleNavigateToDemand = () => {
+    navigate("/demand/create");
+  };
+
   return (
     <MainRoot title="Dashboard">
       <div className="grid size-full grid-cols-3 grid-rows-6 gap-4">
-        <Card className="flex">
-          <CardHeader className="flex flex-row">
-            <TbCalendarClock className="size-14 text-indigo-500 opacity-75" />
+        <Card className="col-start-3 row-span-3 row-start-1">
+          <CardHeader className="text-gray-900 dark:text-gray-300">
+            <CardTitle className="flex flex-wrap justify-between gap-2 text-xl">
+              <div className="flex flex-wrap items-center gap-4">
+                <BsFillInfoSquareFill />
+                <span>Soldes</span>
+              </div>
+              <Button variant="callToAction" onClick={handleNavigateToDemand}>
+                <PlusIcon className="mr-2 size-4" />
+                Nouvelle demande
+              </Button>
+            </CardTitle>
           </CardHeader>
-
-          <CardTitle className="p-4">
-            <span className="text-base">Solde de congés</span>
-            <div className="text-5xl">{currentUser.ca}</div>
-          </CardTitle>
+          <CardContent className="divide-y divide-slate-300 dark:divide-slate-700">
+            <FieldRow title="Congés">{currentUser.ca} jours</FieldRow>
+            <FieldRow title="RTT">{currentUser.rtt} jours</FieldRow>
+            <FieldRow title="Télétravail">{currentUser.tt} jours</FieldRow>
+          </CardContent>
         </Card>
-
-        <Card className="flex">
-          <CardHeader className="flex flex-row">
-            <TbCalendarRepeat className="size-14 text-red-500 opacity-75" />
-          </CardHeader>
-
-          <CardTitle className="p-4">
-            <span className="text-base">Solde de RTT</span>
-            <div className="text-5xl">{currentUser.rtt}</div>
-          </CardTitle>
-        </Card>
-
-        <Card className="flex">
-          <CardHeader className="flex flex-row">
-            <MdOutlineLaptop className="size-14 text-orange-500 opacity-75" />
-          </CardHeader>
-
-          <CardTitle className="p-4">
-            <span className="text-base">Solde de télétravail</span>
-            <div className="text-5xl">{currentUser.tt}</div>
-          </CardTitle>
-        </Card>
-        <Card className="row-span-5">
-          <CardHeader>
-            <CardTitle>Mon service</CardTitle>
+        <div className="col-span-2 col-start-1 row-span-2 row-start-1">
+          <NotificationsCard />
+        </div>
+        <Card className="row-span-4 row-start-3">
+          <CardHeader className="text-gray-900 dark:text-gray-300">
+            <CardTitle className="flex flex-wrap justify-between gap-2 text-xl">
+              <div className="flex flex-wrap items-center gap-4">
+                <GrGroup />
+                <span>Mon service</span>
+              </div>
+            </CardTitle>
           </CardHeader>
         </Card>
-        <Card className="row-span-5">
-          <CardHeader>
-            <CardTitle>Mon équipe</CardTitle>
+        <Card className="row-span-4 row-start-3">
+          <CardHeader className="text-gray-900 dark:text-gray-300">
+            <CardTitle className="flex flex-wrap justify-between gap-2 text-xl">
+              <div className="flex flex-wrap items-center gap-4">
+                <MdHomeRepairService />
+                <span>Mon équipe</span>
+              </div>
+            </CardTitle>
           </CardHeader>
         </Card>
         <MapContainer
-          className="row-span-5 size-full rounded-xl"
+          className="col-start-3 row-span-3 row-start-4 size-full rounded-xl"
           center={[49.10668200684202, 6.17647682854827]}
           zoom={6}
           scrollWheelZoom={true}
